move database interface files into their own module
