+++ src/dolt_annex/gallery_dl_plugin/__init__.py
@@ -21,7 +21,7 @@ import gallery_dl
 from dolt_annex.datatypes.config import Config
 from dolt_annex.datatypes.repo import Repo
 from dolt_annex.datatypes.table import DatasetSchema, FileTableSchema
-from dolt_annex.table import DatabaseConnection, Dataset, RepoDataset
+from dolt_annex.replicated_db.dolt import DatabaseConnection, Dataset, RepoDataset
 
 config_path = Path(__file__).parent / "gallery_dl_config.json"
 skip_db_path = Path(__file__).parent / "skip.sqlite3"
